+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.12.0
+
+* Add a `subscriptionTransformer()` function to create `StreamTransformer`s that
+  modify the behavior of subscriptions to a stream.
+
 ## 1.11.3
 
 * Fix strong-mode warning against the signature of Future.then